pyproject.toml
[tool.poetry]
name = "ai-any-text-clusterer"
version = "0.1.0"
description = ""
authors = ["stevechan <chanshunli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.2"
requests = "^2.32.3"
matplotlib = "^3.9.2"
faiss-cpu = "^1.9.0"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
